fix: dt.Device and dt.Claim.Device conflict
Python supports using the same name for classes nested under different
namespaces, we discovered that our tools for generating documentation
were not able to resolve the difference. This conflict has been fixed by
prefixing both `dt.Claim.Kit` and `dt.Claim.Device` with `Claim`.

While this is technically a breaking change, we came to the conclusion
that due to discovering this problem within an 30min of initial release,
the update will be pushed as a patch regardless. We hope this is
understandable.